clear up code with edge case

+++ metaquery/__main__.py
@@ -20,6 +20,7 @@ def launch_job(args):
         run_R_analyses(args)
         report_job(args)
         export(args)
+    return flag
 
 
 def main():
@@ -45,7 +46,7 @@ def main():
         print("Start time: %s for JOB ID %s" % (time.strftime("%Y-%m-%d %H:%M:%S"), job_id))
 
         query_fp = _args.query
-        outdir = f"{base_dir}/{job_id}"
+        outdir = f"{base_dir}/metaquery_output_{job_id}"
         tempdir = f"{outdir}/temp"
         form["outdir"] = outdir
         form["tempdir"] = tempdir
@@ -60,15 +61,20 @@ def main():
         command(f"cp {query_fp} {outdir}/query.fasta")
         command(f"mkdir -p {tempdir}")
 
-        launch_job(args)
-        command(f"rm -rf {tempdir}")
+        flag = launch_job(args)
+        if flag == "EMPTY":
+            print("EMPTY BLAST RESULTS")
+        if flag == "DONE":
+            command(f"rm -rf {tempdir}")
+            command(f"rm -f {outdir}/job.yaml")
+            command(f"rm -f {outdir}/query.fasta")
         print("End time: %s for JOB ID %s\n" % (time.strftime("%Y-%m-%d %H:%M:%S"), form["job_id"]))
     elif _args.search_by_name:
         search_term1 = _args.search_name.replace(" ", "_")
         if job_id == "":
             search_hash = md5(search_term1.encode('utf-8')).hexdigest()[-6:]
             job_id = f"{search_term1}_{search_hash}"
-        outdir = f"{base_dir}/{job_id}"
+        outdir = f"{base_dir}/metaquery_output_{job_id}"
         if os.path.exists(outdir):
             print(f"WARNING: {outdir} exists. DELETE.")
             command(f"rm -rf {outdir}")
@@ -76,7 +82,9 @@ def main():
         args = {"job_id": job_id, "search_name": _args.search_name}
         args["outdir"] = outdir
         print(args)
-        report_by_name(args)
+        flag = report_by_name(args)
+        if flag == "EMPTY":
+            print("EMPTY BLAST RESULTS")
 
 
 if __name__ == "__main__":
